H 08

+++ services/cloudService.js
@@ -187,15 +187,39 @@ class CloudService {
   }
 
   async deleteInterestPoint(interestPoint) {
-    const interestPointQuerySnapshot = await getDocs(this.interestPointsCollection);
-    const interestPointDoc = interestPointQuerySnapshot.docs.find(
-      doc => doc.data().creator === interestPoint.creator && doc.data().name === interestPoint.name,
-    );
-    if (interestPointDoc) {
-      await deleteDoc(interestPointDoc.ref);
-      return true;
+    const netInfo = await NetInfo.fetch();
+    const isConnected = netInfo.isConnected;
+
+    if (isConnected) {
+      try {
+        const interestPointQuerySnapshot = await getDocs(this.interestPointsCollection);
+        const interestPointDoc = interestPointQuerySnapshot.docs.find(
+          doc => doc.data().creator === interestPoint.creator && doc.data().name === interestPoint.name,
+        );
+        if (interestPointDoc) {
+          // Eliminar de BBDD
+          await deleteDoc(interestPointDoc.ref);
+        }
+        // Eliminar objeto de listado en almacenamiento local
+        let interestPoints = await AsyncStorage.getItem('interestPoints');
+        interestPoints = interestPoints ? JSON.parse(interestPoints) : [];
+        const updatedInterestPoints = interestPoints.filter(
+          ip => ip.name !== interestPoint.name,
+        );
+        
+        await AsyncStorage.setItem(
+          'interestPoints',
+          JSON.stringify(updatedInterestPoints),
+        );
+        return true;
+      } catch (error) {
+        throw error;
+      }
+    } else {
+      const error = new Error('NoInetConection');
+      error.code = 'NoInetConection';
+      throw error;
     }
-    return false;
   }
 
   async clearCollection(collectionName) {
